Add new filters and fields to reservation unit admin panel
